+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ## New Features
 - _Remote Models_ module: UI and server functions for selecting a remote (or local) model and
-obtaining the remote path on github (or the local path) to the selected model
+obtaining the remote path on github (or the local path) to the file of the selected model
 
 ## Version 23.03.1